Drop uninit handling in arrays
Summary:
We have code that checks, at runtime, that new array values in appends and sets are initialized.

As I learned from ricklavoie, these checks are unnecessary for member ops. It's not possible for the RHS of a member instruction to yield Uninit. That leaves only C++ extension code. Unfortunately, such code really does abuse Uninit values - both our own reflection extension and some legacy extensions do. I wouldn't be surprised if there are other cases.

We make "value is initialized" a precondition of the appendMove / setMove methods on the bare ArrayData, but we do NOT make it a precondition on array-init and type-array helpers. These helpers use tvToInit instead. This way, we keep common member ops as tight as possible while also being robust to weird extension code. We assert at this boundary in debug mode.

Reviewed By: ricklavoie

Differential Revision: D27642303

fbshipit-source-id: bc0bc458376bca6fbde3224cc2f5cbc42a8adab4