Make append/set IR ops take InitCell
Summary:
Now that the array setters themselves have a "initialized" precondition, we should reflect that in the JIT.

To do so, we need to make use of flavor invariants in irgen code. Luckily, translateInstr already asserts the types of stack locations based on these environments, so all we need to do is strengthen popC and topC to assert TInitCell and fix up a couple places where we should be calling popCU instead. (In irgen-inlining, for example, the stack values are parameters we want to pass to a callee and default params are Uninit, so we must use popCU.)

Reviewed By: ricklavoie

Differential Revision: D27651845

fbshipit-source-id: 69af6188ddb3ad17d3ac4f2569cb06923e1078c4