Separate the kinds of type constants in the AST
Summary:
Type constants have 3 distinct kinds

- abstract type constants
```
abstract const type T [as Tupper] [super Tlower] [= Tdefault];
```
- concrete type constants
```
const type T = Tval;
```
- partially abstract type constants
```
const type T as Tupper = Tval;
```

However, they are represented by a single bag of optional values. This leads to strange patterns where a type constant would be judged abstract based on the presence of a constraint, or concreteness based on the presence of `=`, which is incorrect for abstract type constants as = is for defaults.

{D14788287 (https://github.com/facebook/hhvm/commit/1fb7ca9ecd0256ad95b5ca4b11fc839a1a956685)} added a field to disambiguate; this extends the separation so that only the expressible kinds of type constants are represented.

For example, a partially abstract type constant must always have a constraint and a value; concrete type constants do not have a constraint; with abstract type constants, both constraints are optional and the default is optional. Code no longer simply checks fields in the AST but rather asserts a specific kind of type constants.

Reviewed By: losvald

Differential Revision: D26626204

fbshipit-source-id: 9ce95e6fcb1afc85831fb0f45f44aa19d96b03bb