Eliminate Pos.none in desugaring
Summary:
Type errors that have `Pos.none` as positions report that "There is an error somewhere in this file. However the type checker reports that the error is in another file".

We use `Pos.none` as placeholders in the desugarer. As a result, there are some cases where we get the above error when we can probably be a bit more specific about where the error arises from.

This diff tries to place better positions in places where we were previously using `Pos.none`

Reviewed By: Wilfred

Differential Revision: D27523604

fbshipit-source-id: 01cd0aebae944d12b56629e2bc30bf7c46d511cd