Use non user denotable variable for visitor
Summary: Currently using `$v` as the variable for our visitor within the desugared lambda. While this variable shouldn't be revealed to the user as there is currently no way to sneak an unbound variable into the desugared lambda, we still may want to keep this sanitary by changing this variable to be `$0v`.

Reviewed By: Wilfred

Differential Revision: D27491087

fbshipit-source-id: b5c182248338ffe241637256f1f664f95d3a16f6