Handle interface methods properly
Summary: We don't include interface methods in the method table for classes that implement them (or interfaces that extend them), the assumption is that eventually they will be implemented and the implementation will be part of the table. This fixes the `meth_caller` checks so that it's still possible to construct a meth_caller using an interface method.

Reviewed By: ricklavoie

Differential Revision: D27659488

fbshipit-source-id: ec2301dfb825ce7b11018c9786c90899a5870af4