Make uncounted dec-ref API match counted API
Summary:
Our refcounted value API separates "dec-ref" and "release". Let's do the same for the uncounted value API. Mixing up these two operations caused the bug that binliu19 fixed in D27040240 (https://github.com/facebook/hhvm/commit/44faa99d1953e566579b1aae9464cc60062b6bc9).

We introduce DecRefUncounted, and the specialized variants DecRefUncountedArray and DecRefUncountedString. We also make releaseUncounted() one of the virtual ArrayData methods, so we don't need to manually switch over array kinds outside of ArrayData implementation files. DecRefUncounted is the uncounted analogue of decRefAndRelease, while ReleaseUncounted is the uncounted analogue of release.

Today, if we uncountedDecRef() an object with a count of UncountedValue ("uncounted refcount 1"), we'll leave it's count at StaticValue. That's dangerous, because it's a valid-but-incorrect count. isStatic() will return true for the value, which could result in us doing a host of weird things with it. To make code safer we introduce an (invalid) refcount value between UncountedValue and StaticValue: UncountedZero. An object with this refcount will quickly fail any invariants checks. There is nothing valid to do with such an object other than to call ReleaseUncounted on it.

In the ReleaseUncounted methods, we'd like to assert that the value has count UncountedZero, but such a value won't pass basic array sanity checks. Instead, we assert that it does not need an "uncounted cow check" - i.e. it's at "uncounted refcount 1". At the place we do the dec-ref, if we go to zero, in debug mode, we fix the count, like we do for counted dec-refs.

Finally, we drop the possibility of "failure" in uncountedIncRef(). Now that we've cleaned up one-bit refcounting, it always succeeds. For places that want to do an uncounted inc-ref on a maybe-counted value, we provide persistentIncRef(), which succeeds for statics, succeeds and inc-refs for uncounted values, and fails for counted values.

Reviewed By: binliu19

Differential Revision: D27721374

fbshipit-source-id: 2e7d702659a52828d280df4c84d83d648a61b1fb