Stop dynamic being potentially considered enforceable
Summary:
Type `dynamic` is not enforced by hhvm, and so it should not be considered to be enforced. Otherwise, there is a risk that `coerce_type` is called with an enforced `dynamic` which will do the wrong thing.

Also, do some light refactoring in preparation for adding support for partially enforced types, and add an .mli file

Reviewed By: andrewjkennedy

Differential Revision: D27651634

fbshipit-source-id: 414248fcb4855684d61d030ae6bbd1f37856d83a