Pull out MakeUncounted array and string helpers
Summary:
Make "makeUncounted" another virtual ArrayData op. In a follow-up, I'm going to replace "onSetEvalScalar" with "makeStatic", at which point we'll have the API we want.

Pull as much code out of the individual array MakeUncounted implementations as possible:
- The array hash lookup moves into a generic helper. Layouts don't need to update seen.
- The string logic is repetitive and is duplicated in ConvertTvToUncounted. Make it a separate helper.
- Updating APC stats moves to the allocation routines so we can't miss doing it.

This diff also restores the ability to make an uncounted bespoke array (not exercised in any way, yet). We now escalate in MakeUncountedArray instead. We don't populate the "seen" map on escalation anymore; instead, we are just robust to missing values in this map in MakeUncountedArray, like we are in MakeUncountedString.

Reviewed By: colavitam

Differential Revision: D27748599

fbshipit-source-id: 59972975f2d2f774d117bed8d7162b3d9c0d7d07