Fix enforceability inheritance behavior
Summary:
The `ttc_enforceable` field on a typeconst indicates whether it can be used in an `is`/`as` expression. In legacy decl, a typeconst is enforceable if it has the `<<__Enforceable>>` attribute or if it overrides some parent typeconst with the attribute. We've never modeled this behavior correctly in shallow decl because it causes a severe performance regression (or at least, a naive early attempt at modeling it in D14528842 did). Since enforceability is propagated down the hierarchy, even through overrides, we're rarely able to stop traversing the linearization early when looking up a typeconst--it's always possible that some ancestor may have declared an enforceable version, so we exhaust the linearization looking for one (and usually don't find one).

We do need to model this behavior, though! I speculated when this behavior was implemented (on D14330721 (https://github.com/facebook/hhvm/commit/d0ed9d7578934555a34784c86e600e96c5a1df60)) that we might be able to remove this overriding behavior, and change the semantics of `ttc_enforceable` to mean "the `<<__Enforceable>>` attribute was shallowly present on the declaration of this typeconst". If we required the `<<__Enforceable>>` attribute to be explicitly present on overrides, then we could make that semantic change to `ttc_enforceable` (i.e., so that it would mean the attribute was shallowly present rather than present-here-or-on-any-ancestor) without changing the overall behavior (i.e., overrides of enforceable typeconsts are still guaranteed to be enforceable). But I gave this a try in D27438472, and it emits nearly 800K errors in WWW. I suppose this is why the override behavior was implemented in the first place--it made the feature easier to roll out without extensive codemods.

This diff, instead of changing the semantics of typeconst canonicity (as D14528842 did), provides a new API for looking up typeconst enforceability. This API is used only when we need to know whether a typeconst is enforceable (i.e., any time we would have read `ttc_enforceable`), dramatically reducing the performance impact of this change. This diff also fixes a bug(?) in legacy decl, where enforceability could be erased when ancestors are declared in a particular order (this behavior is demonstrated in the test case added in the previous diff, implement_interface_abstract3.php).

Reviewed By: vassilmladenov

Differential Revision: D27524511

fbshipit-source-id: a4e951ca71aa12f7c6cbe3637cfd660b2fc3e1b4