Improve the treatment of various Collections with dynamic
Summary:
Add logic to take advantage of hhvm's enforcement of collection types with `dynamic`.

If a function parameter or return has a collection type, hhvm will enforce that the runtime value is of that collection; however, it will not enforce that the contents are of the given type. Hence, when an argument of type `dynamic` is passed to a `vec<t>` parameter, and the check succeeds, we know that it is a `vec`. We also know that the contents are subtypes of `dynamic`, because that's the only way for a `vec` to subtype `dynamic`. Thus, the call can be allowed as long as `dynamic` is a subtype of `t`. Similar reasoning applies to non-covariant collections such as `Vector`, except that `t` must be also be a subtype of `dynamic` (due to invariance).

It is also safe to treat `Container`, `KeyedContainer`, and `Traversable` similarly, even though these interfaces are not subtypes of dynamic. For the first two, the only possible implementors of the interfaces are `vec`, `Vector`, and the other collections, so the contents must subtype dynamic. There is no such guarantee for `Traversable`, since there is no restriction on implementing its `Iterator` sub-interface. However, for any class that implements `Iterator` and is also a subclass of `dynamic`, the extra checks required by `<<__SoundDynamicCallable>>` on the `current` method ensure that it only returns subtypes of `dynamic`.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D27696205

fbshipit-source-id: e142218aa19421e30337f393e447e8a3190aa582