Avoid dead code in ifThen() in profiledType() by eliminating unnecessary guard
Summary:
ifThen() is unable to properly propagate frame state to an unreachable branch
and when we try to emit code for such branch, we may use an incorrect frame
state. This is problematic if both branches try to pop things out of the stack,
as is the case with implDictKeysetIdx() using profiledType(), where if the
profiled type is impossible, the last finish() in implDictKeysetIdx() will be
dead and will start with an empty stack state, trying to pop things again,
causing assertion errors.

Fix this particular instance by avoiding the guard that is always true or
always false. This is worth doing anyway, as we should not exit translation
if the profiling data were unusable.

A generic fix of ifThen() + ifThenElse() would require a larger audit of all
use cases, as in this case the unreachable code was outside of the ifThen().

Reviewed By: oulgen

Differential Revision: D27802133

fbshipit-source-id: e261312d8d51448b8e52ed231d6cd5bd2058923f