Tweak array-related APC invariants
Summary:
It's difficult to maintain certain array-related APC invariants while making the storage more flexible. In this diff, we swap one invariant for another.

The invariant we drop is that, if we get to APCArray::MakeUncountedArray, we'll co-allocate an APCTypedValue right before that array. This invariant prevents us from re-using persistent arrays, even if we can detect them. We attempt to correct for this case in APCTypedValue::HandlePersistent, but when bespokes are involved, we can still fish out these persistent arrays later, e.g. upon escalation of a LoggingArray. We solve this issue by making this optimization a pure request - no guarantee of success - and checking for it on construction.

The invariant we add is that ConvertTvToUncounted *never* leaves a bespoke array in the result. We intended to obey this invariant anyway, but there were a few places where we didn't enforce it. The new code satisfies the invariant by making two changes: first off, with a check in APCTypedValue::HandlePersistent, which handles top-level arrays, and second, by fixing MakeUncountedArray.

Reviewed By: ricklavoie

Differential Revision: D27782763

fbshipit-source-id: 2598b5684e6adcdb3b9336be8d7aa7494b0f6892