Replace FPInvOffset (frame relative) with SBInvOffset (stack base relative)
Summary:
FPInvOffset has very confusing semantics. These offsets are already not frame
relative in resumables, where frame lives on the heap, and neither in
prologues, where the location of the frame is unknown.

Replace it with SBInvOffset, which is relative to the stack base. In regular
execution contexts it is located Func::numSlotsInFrame() away from the previous
FPInvOffset, in resumables and prologues it is the same value as before.

Reviewed By: ricklavoie

Differential Revision: D27583348

fbshipit-source-id: 43225f8a27217aa6207e029ac3848f6a58fcfbf4