Make schema validation more efficient
Summary:
We should only do these checks once per schema.

Note that this change is a slight behavior change, as we'll validate the serialization spec even if we get, say, an empty map. I think the extra validation here is arguably correct - it's bogus to deserialize an empty map if the Thrift spec doesn't have a key or val type - but if we want to preserve the current behavior, we can just add a size check around the O(1) validation. I'll let Thrift folks make that call.

Reviewed By: yfeldblum

Differential Revision: D27824216

fbshipit-source-id: 36c0046773db493152c0854163fa3151e6824e43