Fix smashing of ProfPrologues' callers when TC fills up
Summary:
If the TC was full before we started to relocate the optimized
translations/prologues of a function, then this functions' translator structures
were just thrown away.  This would prevent the callers of the profiling
prologues of such functions from being smashed, so they would continue to call
profile code, which would cause a crash once the prof TC code block was freed.
Instead of throwing away the translators, this diff resets their transMeta, thus
signaling that the translator failed and allowing callers to be smashed to stop
calling profile code.

Reviewed By: ricklavoie

Differential Revision: D27816239

fbshipit-source-id: 35dd5c4ce7b0f08810c4d9339cd17d88dda8e55f