Lazily guard locals and stack locations in the tracelet selector, take 2
Summary:
This diff relands D27545816 (https://github.com/facebook/hhvm/commit/3ecba8ccfdc1b7dce556ac6a806036b16b95bd04). The reason that diff was reverted is because it
exposed a latent JIT bug by making JITing profile and live translations
significantly faster.  This caused live translations to fill up code.main before
retranslate-all had finished producing optimized translations.  The code.hot
filled up and the JIT ran out of space to place the optimized translations.  The
bug was that the callers of the corresponding profile translations were not
being smashed to stop calling profile code in this case. D27816239 fixes that
bug.

Original diff summary:

This diff reworks the tracelet selector's mechanism to compute the guards
needed.  Previously, guards were eagerly inserted for all live VM locations
before processing any bytecode instruction, and then guard relaxation would
eliminate the unnecessary guards.  Although simple, that approach resulted in an
unnecessarily large number of locals and stack locations being tracked in
FrameState, leading to significant JIT overheads for forming regions in
functions with too many locals and/or stack slots used.  This was particularly
wasteful because the tracelet selector has a limit on the number of locations
that it will guard (currently 5) before breaking the region.
This diff switches to a lazy guarding approach to reduce the JIT overhead in
such cases.  With lazy guarding, new guards are added on demand, right before
the VM location is accessed, thus drastically reducing the number locations that
need to be tracked in FrameState for tracelet selection.  To benefit from the
fact that not all locations need to be tracked in FrameState, the vectors of
locals and stack slots in FrameState were replaced with hash maps to allow a
more efficient sparse representation.
This diff also adds a runtime option to allow eagerly guarding regions up to a
specified number of live VM locations.  This limit is set to 0 by default
(meaning, always lazy guarding), but it can be used to partially or completely
disable lazy guarding.

Reviewed By: ricklavoie

Differential Revision: D27822549

fbshipit-source-id: 720126f76c184e802aecda0969c56629ddbbdc32