Add CheckROCOW to ReadOnlyOp immediate
Summary:
Precursor to supporting modifying readonly value type arrays.

Supporting this feature will require different runtime enforcement on member op chains, specifically for the last prop access, and the following element accesses (in write mode).

According to the bytecode spec,  `ReadOnlyOp::Mutable` means HHVM will enforce that this prop is mutable, and similarly for `ReadOnlyOp::ReadOnly`.

Since the new check will be "Can be readonly, if value type. Otherwise must be Mutable", it makes sense to make this a new immediate, as it doesn't fit under the existing rules.

[RFC] Where best to set this
To set this, I was thinking HackC could emit it. It appears HackC already loops over the member op chains when constructing them in a way that would make it easy to add some additional logic to resolve the last one. Alternatively, we could add some logic in hhvm to scan the bytecode as well.

Reviewed By: ricklavoie

Differential Revision: D27861665

fbshipit-source-id: c2492aa59150333dfe5fce44c3d0cbb9c1614272