Flag to ban definitions for partially abstract type constants
Summary: The prior flag for partially abstract type constants (`tco_disable_partially_abstract_typeconsts`) disables the ability to override them. This flag errors at the definition site. The prior flag is useful to discover sites that override PATCs, but in practice it ends up being more useful to eliminate the definitions first, as ones that are never overridden can simply be made concrete

Reviewed By: Wilfred

Differential Revision: D27864041

fbshipit-source-id: dd2ae33f4d2db1e9fe1c2851e1dd4edbd72ab035