Add env variable to MakeUncounted helpers
Summary:
We're going to need to plumb a couple more parameters through these functions. At a bare minimum, we need to tell them whether to escalate vanilla arrays or not, but after we get a basic version working, we might want additional flags so we can interleave layout checks instead.

It's annoying to plumb these parameters right now. Introduce a param struct to make it easy to do so. Yes, that puts "seen" at one level of indirection further than it is today, but for cold code like this it won't matter.

Reviewed By: colavitam

Differential Revision: D27891427

fbshipit-source-id: e8ba91ed09e07401027cedf42fc0c56711368095