Parser option for brace escape
Summary:
Adds parser option for making \{ a valid escape sequence

So that
```
$x = 5;
$y = 6;

Currently:

var_dump("\{$x,$y}") => "\{5,6}"
var_dump("{$x,$y}") => fatal

We would like:

var_dump("\{$x,$y}") => "{5,6}"
```

This was a task I had as an intern and didn't end up getting to. Found an old diff relating to it and decided to finish it off.

Reviewed By: shayne-fletcher

Differential Revision: D27776520

fbshipit-source-id: 111098806a66ed49328d0b7db68a2e05c6f595d4