Add benchmark for hackc
Summary:
Walks an entire directory tree collecting all `.php` files loading their contents into memory and computing their hashes, then in parallel compiles everything, recording the wall time and CPU time as well as peak and final RSS and files which failed to compile.

This tool may also be useful for anyone that wants to do a quick scan of www with an hhvm change and see if anything is failing to compile.

Reviewed By: shayne-fletcher

Differential Revision: D26908074

fbshipit-source-id: 44e5806901db4ee27f218aa0bb511cd941fe07b5