Add logging for file tearing during requests
Summary: When decls in compilation is introduced we may have to handle situations where bytecode is requested for a file that was changed on disk after decls were already consumed as a dependency of an already loaded file. We either need to detect this situation and terminate the request or handle it gracefully by recovering the bytecode based on the prior version of the file. Terminating the request will be far simpler, therefore we're hoping to measure how often this happens in practice.

Reviewed By: ricklavoie

Differential Revision: D27805653

fbshipit-source-id: 48edeeedb19cce16b21ad324f7b1e69842204113