legacy decl parser should pick up scc_refs from FunctionPointers
Summary:
```
const mixed X = S::F;
const mixed Y = S::f();
const mixed Z = S::f<>;
```
What are the "scc_refs" that we need to track for cyclic initializers?
* X: does need to track S::F, and legacy-decl does, and direct-decl does
* Y: arguable whether it should track S::f; legacy-decl does and direct-decl does
* Z: doesn't need to track S::f; legacy-decl doesn't but direct-decl does

We're not going to change the way the direct-decl-parser works for "Z". Our goal of keeping it slim and lightweight is more important than correctness.

Therefore, this diff changes the legacy-decl parser for "Z" to bring its behavior in accord with the direct-decl-parser.

Reviewed By: vsiles

Differential Revision: D27836472

fbshipit-source-id: 8bf632a51af00e53af7d87c7553debb5048a1de8