legacy decl parser should pick up scc_refs from shape fields
Summary:
The field "scc_refs" is an attempt to detect some cases of cyclic references during const initializers; they cause fatals in HHVM the first time you access the const.

How is the scc_refs field populated? There's one path in the direct-decl-parser, and another path in the legacy-parser.

Consider the case of shape keys:
```
const mixed A = shape(TestE::X => 1);
```
If there's a cycle involving TestE::X then HHVM will fatal when you access "A". So, it makes sense that cycle-detection scc_refs should include TestE::X in this case.

The direct-decl-parser does indeed add TestE::X to scc_refs in this case, so detecting a cycle. The legacy parser doesn't.

This diff fixes the legacy parser to detect a cycle.

NOTE: the way the code is written, the legacy parser has apparently already resolved "self::" references by the time it gets here, so it's not able to distinguish self:: shape keys from the fully-written-out form. I don't intend to fix that here.

Reviewed By: vsiles

Differential Revision: D27829300

fbshipit-source-id: 5ca49e574e9db09881155dbf8830e2e9b0e44d87