add begin_const methods
Summary:
Currently the decl parser gathers references of the form "enum member E::X depends upon some other constant S::R". It does this to detect some cases of circular initialization.

Arguably hack static analysis has no business doing this, and it should be done by lints, but that's a debate for a future day.

Anyway, at the moment the codebase gathers them by constructing nodes for expressions, and then visiting them to extract references. This seems wasteful.

This diff instead introduces a new way to handle it. At the start of parsing an enum member we turn on "reference accumulating". Any time we encounter a reference "S::R" then we'll add it to the accumulator. Once we're done parsing the enum member, well, we've got a complete list of references. No need for a subsequent visitor. No need to dive into expressions.

Does this feel like a hack? I think no it doesn't. It's pretty much a standard way for folks to write actions in ANTLR grammars, e.g.
```
decl: 'const' NAME '=' {ActionConstStart} expr ';' {ActionConstEnd}
```

Anyway, what this diff does is add three "start actions" for the three places where we'll want to turn on reference accumulating.

Differential Revision: D27808015

fbshipit-source-id: e86c8e05ed94a1e03b9dfd16f5f4a80adc8a7ff4