Replace RepoAuthoritative sqlite repo with RepoFile
Summary:
Hook up RepoFile everywhere to replace the usage of the sqlite repo in
RepoAuthoritative mode. The legacy Repo still remains, because it can
still be technically used in sandbox mode (though that will be removed
soon). Most of these changes are straight-forward replacements of Repo
interfaces with equivalent RepoFile interfaces.

Use instantiations of TokenOrPtr (inside LockFreePtrWrapper) in Func
to deal with lazy loading of bytecode or line tables. Note: lazy
loading is now only supported in RepoAuthoritative mode, and therefore
only for RepoFile. In sandbox mode, if you're using Repo for some
reason, we'll now always eagerly load the data.

Things are set up so that RepoFile uses the existing config
transparently. So, this change should just work with no needed change
in config. In the future I plan to remove most of the Repo options and
add a set tailored just for RepoFile.

I removed support for local disk caching for RepoAuthoritative builds
in compiler/. It can be supported, but we've never really used it, and
it adds unnecessary complexity (plus I have a better solution).

Loading from RepoFile is significantly faster than Repo, as is
writing. This is most obvious with HHBBC.

Reviewed By: jano

Differential Revision: D27876896

fbshipit-source-id: 503e93adb40470aea56a828d64221559ef250ff2