Allow for parallel encoding of UnitEmitters
Summary:
Encoding UnitEmitters can be the bottleneck during repo-auth builds
(writing to disk is relatively inexpensive). So, rather than force the
encoding and write to be done in the same call, allow for separate
encoding of the UnitEmitters. Multiple threads can pre-encode the
UnitEmitters. Adding them to RepoFileBuilder (thus writing them to
disk) still must be done from a single thread.

During a repo-auth build, instead of maintaining queues of
UnitEmitters, maintain a queue of EncodedUEs instead. For certain
options, we still need UnitEmitters (for dumping hhas), so optionally
store those as well.

Differential Revision: D27897490

fbshipit-source-id: e272d23f6bfb190d6f3e7a489450dbfe740dcf4f