Add (but not use) RepoFile
Summary:
In an effort to make review easier, I'm splitting up the introduction
of RepoFile with the changes necessary to actually use it. Therefore,
this diff merely adds RepoFile (and a few other changes necessary to
make everything compile).

RepoFile is the new replacement for the sqlite repo. It only supports
RepoAuthoritative mode and therefore does not need to support
arbitrary insertions. The file format is designed to be lightweight
and (most importantly) very quick to generate. A repo file is created
via a RepoFileBuilder, and once created, is immutable. Unlike the
sqlite repo, it does not support arbitrary queries and is instead
designed to support only the specific queries we need.

Reviewed By: jano

Differential Revision: D27870483

fbshipit-source-id: e2a1ae722a8bf74c90b3d17f8f63edc6e60bf923