Add TokenOrPtr
Summary:
For lazy loading various information from the repo, we want to store
either a pointer to some data (if its been loaded), or a
RepoFile::Token (which tells us where to retrieve it from). TokenOrPtr
is an utility class which accomplishes this. It can store a pointer to
T, or a RepoFile::Token compactly in 64-bits (along with the necessary
bit to distinguish the two cases). In addition, it keeps the lower 2
bits free, so it can be used in a LockFreePtrWrapper.

Reviewed By: paulbiss

Differential Revision: D27876448

fbshipit-source-id: 273b9a2f5caf93f0450ed30ea06842185d4dabc4