add test baseline for return positions
Summary:
The legacy decl parser gets some positions wrong. I haven't yet investigated why. This diff merely establishes current behavior in a test.

The specific problem is that if the return type of a function/method is `<<__Soft>> varray_or_darray<T>` or `<<__Soft>> vec_or_dict<T>` then the decl records the position of the return type as including the `<<__Soft>>` attribute. It only does so for these two specific types. It doesn't do it for others like darray, nor bool, or anything else.

Reviewed By: vsiles

Differential Revision: D27837851

fbshipit-source-id: 252e46fdd6c919bd6cc152555cdbd823718cb445