Verifier support for CheckROCOW
Summary:
CheckROCOW must only appear on the last property access in a member op chain.

- `CheckROCOW` can only appear on prop-flavored dims + BaseSC
- `CheckROCOW` can only appear on the *last* prop-flavored dim

We'd like to also enforce that it *must* appear if there are bytecodes that support it in the member op chain. We can't add this check yet because HackC support isn't finished, so currently all the immediates are `ReadOnlyOp::Any`. Filed a task: T89353653

Reviewed By: ricklavoie

Differential Revision: D27905533

fbshipit-source-id: ab44eea180268fd95486cd0b84ca5892af5efb84