Support writing to roProp bit in MInstrState
Summary:
This diffs allows us to support modifying readonly value type collections.
Consider:
```
$foo->x[1][2][3]->p[1][2][3]->a[1] = $x
```
In a readonly world, we need to ensure none of the accesses on the lhs are modifying readonly values. Currently, we enforce a "must be mutable" on each of those member ops. However, if these accesses are on arrays with value types, modifying them does not violate readonly semantics.
In order to support this, we have determined the general rule to be: **The last prop access may be readonly if its a value type, and then all following element accesses must also be value types. The rest of the prop/elem accesses still must be mutable.**

In order to support this in the runtime, we just need single bit in the MInstrState as to whether the last property access in a member chain is "readonly && value type". The element accesses can read this and determine whether they must be value types.

We want to only write for the last property access. We want to read for all elem accesses after this last property access.

- Specialized type to `PtrToMISBool` so we can more easily identify this ptr in memory-effects and more precisely determine when we actually write.

Reviewed By: ricklavoie

Differential Revision: D27891682

fbshipit-source-id: 7e676850a30b4550714e38c9982d6403b3a1c57a