Fix skipifs and broken tests in hphp/test/zend/good/ext/ldap/tests
Summary: All 77 of these tests were skipping due to top-level code in skipifs. Fix the skipifs (by migrating the check for the ldap extension that all the tests want into a config.runif, and fixing the remaining ones) , as well as a number of problems in the tests themselves that were hidden by the skipping.

Reviewed By: ricklavoie

Differential Revision: D27938047

fbshipit-source-id: 251732c7002739868782932edebd61c6c370af5b