Fix skipifs and broken tests of the 'standard' extension.
Summary:
All but one of the .skipifs in this directory had top level code. Convert them all to runifs, which fixes the ones that had top-level code.

A number of the tests were broken and also needed fixing. We had top level code and test expectations that needed to be adjusted post HAM unification.

Differential Revision: D27938036

fbshipit-source-id: 5e146078384e02b937735b8469bf8a118035cb80