Fix skipifs and broken tests of the fribidi extension
Summary:
hphp/test/slow/ext_fribidi/config.runif had top-level code. Fix it by converting to a config.runif

All the .expect files in this directory needed to be adjusted for HAM unification.

hphp/test/slow/ext_fribidi/002.php.expect needed substantial changes, but as it notes on line 41 the output there is subject to change (presumably with library version updates).

hphp/test/slow/ext_fribidi/004.php needed a partial rewrite. It appears to have been testing passing invalid arguments to fribidi_log2vis and to have been assuming that it would just get a result back. When the argument's type is wrong we raise a fatal, so the test can't continue to work as is. Narrow it down to just testing the empty string input, we don't need to test the argument type enforcement machinery here.

Differential Revision: D27938046

fbshipit-source-id: 614366d6639fe34605701266ebee913132bdc067